refactor : 코드리뷰 후 리팩토링
- SecurityConfig :: 요청에 대한 접근 설정
- CommonResponseDTO :: 겹치는 dto 삭제
