+++ src/main/java/com/example/hobbybungae/HobbyBungaeApplication.java
@@ -1,4 +1,4 @@
-package com.example.hobbybungae.domain;
+package com.example.hobbybungae;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
