--- src/main/java/com/example/hobbybungae/domain/dto/CommonResponseDto.java
@@ -1,10 +1,0 @@
-package com.example.hobbybungae.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public abstract class CommonResponseDto {
-    private String status;
-}
