--- src/main/java/com/example/hobbybungae/domain/hobby/CommonResponseDto.java
@@ -1,4 +1,0 @@
-package com.example.hobbybungae.domain.hobby;
-
-public class CommonResponseDto {
-}
