+++ src/main/java/com/example/hobbybungae/domain/user/UserController.java
@@ -1,6 +1,6 @@
 package com.example.hobbybungae.domain.user;
 
-import com.example.hobbybungae.domain.dto.CommonResponseDto;
+import com.example.hobbybungae.response.CommonResponseDto;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
-    
+
     @PostMapping
     public ResponseEntity<CommonResponseDto> signUp(@RequestBody @Valid UserRequestDto requestDto) {
         return userService.signUp(requestDto);
