+++ src/main/java/com/example/hobbybungae/domain/user/UserService.java
@@ -1,8 +1,8 @@
 package com.example.hobbybungae.domain.user;
 
-import com.example.hobbybungae.domain.dto.CommonResponseDto;
-import com.example.hobbybungae.domain.dto.ErrorResponseDto;
-import com.example.hobbybungae.domain.dto.SuccessResponseDto;
+import com.example.hobbybungae.response.CommonResponseDto;
+import com.example.hobbybungae.response.ErrorResponseDto;
+import com.example.hobbybungae.response.SuccessResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+//@Transactional(readOnly = true)
 public class UserService {
 
     private static final String SIGNUP_ERROR_MESSAGE = "중복되는 회원 아이디가 존재합니다. 다른 아이디로 시도해주세요.";
