+++ src/main/java/com/example/hobbybungae/response/CommonResponseDto.java
@@ -1,10 +1,10 @@
-package com.example.hobbybungae.domain;
+package com.example.hobbybungae.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CommonResponseDTO {
+public abstract class CommonResponseDto {
     private String status;
 }
