+++ src/main/java/com/example/hobbybungae/response/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.hobbybungae.domain.dto;
+package com.example.hobbybungae.response;
 
 import lombok.Getter;
